Revert "Changed Token Operand from int to double to allow proper division results"
This reverts commit 63772a116063ec0a7154bf43bc89fa57b9cc9673.

+++ luirary/src/Driver.java
@@ -78,7 +78,7 @@ public class Driver
             equation.printEq();
             System.out.println();
 
-            double result = equation.evaluatePrefix();
+            int result = equation.evaluatePrefix();
 
             if (equation.hasError())
             {
